mia: improve CIC detection algorithm
Currently, we detect the CIC to emulate by calculating a CRC32 over
IPL3 bootcode in ROM, and keeping a database of the known IPL3s. This
works because right now there are only a handful of IPL3 variants
(exactly one per CIC).

In preparation for libdragon releasing an open source IPL3 that could
evolve and be forked an unbounded number of times, we must stop keeping
a database of IPL3s.

This commit changes strategy: it simulates the IPL2 checksum hash over
IPL3 (which is exactly what a real N64 does at boot) and verifies
if the checksum matches that expected by any of the known CIC variants.
This works because open source IPL3 variants must anyway have a checksum
that collides with that of CICs (which must be obtained via a second
preimage attack, bruteforced on GPUs), otherwise they would not work
on real hardware nor on Ares itself (which correctly simulates the boot
sequence and would refuse to boot an "unsigned" IPL3).