n64: simulate SysAD freezes when accessing non-RDRAM areas via cache
The CPU will freeze until power-off if a cached access is performed to
a non RDRAM area. This obviously never happens with games, but it is
a not so uncommon error when doing very low level development on the
system.

This patch emulates the freeze for the data cache, and for the instruction
cache when using the interpreter. Implementation for instruction cache
and recompiler is left as a followup.