fc: use hkrom for mmc6
MMC3 and MMC6 share the same mapper in the iNES format. To disambiguate
MMC6 from the MMC3 mapper the NES 2.0 submapper 004:1 should be used:

https://www.nesdev.org/wiki/MMC6

Code for using the MMC6 was actually commented in this file up to commit
bed4ce3659a92e, where it was deleted. This commit reintroduces this
code to use it with the NES 2.0 submapper 004:1. This adds compatibility
with the games "StarTropics" and "Zoda's Revenge: StarTropics II".