n64: improve RSP DMA cache coherency check
It is quite normal for RSP ucode to fetch extra data to IMEM/DMEM.
For instance, games tend to load 4 KiB of ucode even if the actual
ucode is smaller; or they can fetch a command buffer in fixed size
chunks of 256 bytes, and then just ignore data past the actual end.

To avoid tons of false positives, we track the actual DMEM cells that
contain tainted data, that is, data read from RDRAM in a non coherent
state. If and only if those DMEM cells are accessed, we issue the
warning.