hiro: prevent exception originating from appkit
As noted in the Chromium source, the value for NSTreatUnknownArgumentsAsOpen
needs to be of type string despite being a boolean setting.

If the type is boolean, AppKit throws an exception (which is caught),
but if unwinding fails (observed on some arm64 systems in certain build
configurations when ares is compiled with LTO and -fomit-frame-pointer)
it terminates the whole application.

I've confirmed the behavior the original commit was designed to correct
is still fixed with this change.