nall: correctly handle TCP peer disconnections in tcp-socket.cpp
Currently, when a client disconnects, the server in Ares fails to
notice and to call the onDisconnect hook, which in turns cause the
server to stop responding (it won't accept a new client, and will
never receive more data from the disconnected one).

This commit fixes it by correctly handling the recv() return code
to detect peer disconnections (return value 0) and any other kind
of error (return value < 0).