n64: clear r0 only when it was written to
Currently a per-instruction epilogue is needed to make sure
the read-only register r0 stays as the value 0. To improve
performance of the generated code, we'd like to move this
operation to be done only when needed.

This commit solves the above problem in two ways: simple
instructions are never emitted if their destination register
is r0, and more complex instructions now emit a clear operation
right after. This way exceptions are still thrown when needed.

The interpreter still clears r0 in the epilogue function.