ruby OSS: use _bufferSize depending on number of channels
The previous implementation used a fixed number of 64 samples in the application
buffer. This value should be a multiple of the number of channels in order to
prevent writing incomplete frames to the audio device. Now the implementation
stores 32 complete frames in the application buffer. This value is derived from
the previous standard configuration of 2-channel stereo audio, i.e
64 samples / 2 channels = 32 frames.