n64: add support for multi-byte writes to gdb
Currently, gdb hooks in n64 doesn't allow for large sizes of writes
because it wouldn't be obvious which data size use on the bus.
While this is true, it doesn't really matter for RDRAM where all data
sizes work in the obvious, correct way. So relax the constraint to
allow writing large portions of RDRAM with a single gdb command.

Co-Authored-By: Chase Meadors <c.ed.mead@gmail.com>