Enforce newlines at end-of-file (#1604)
Enforce newlines at end-of-file in the codebase.

While every modern compiler is able to deal with not having EOF
newlines, it is still a good idea to enforce this.

ParaLLEl-RDP has a few files without EOF newlines, so I surrounded the
n64 Vulkan `include` with a diagnostic pragma to ignore it. This avoids
having to deal with compiler errors if ParaLLEl-RDP is updated, pending
newlines added upstream.