Stage resources into the output directory on Linux.
Originally, to run ares on Linux after building it, you'd `make install` which
copied everything into the right places in your home directory. Windows and
macOS don't have the same traditions about installing apps, so on those
platforms the makefile staged a "portable" installation in the output directory.

Since then, `make install` was modified to install ares system-wide. It's
possible to use it to install ares into your home directory, but a bit overkill.
For consistency, let's also stage a "portable" installation in the output
directory on Linux and the BSDs, so it's easier to test on those platforms.